build: add explicit dependencies for guava and protobuf (#10)
* build: add explicit dependencies for guava and protobuf

Dependency checker identified guava and protobuf as used dependencies
that were not explicitly declared in pom.xml. I have added them and
updated the version of guava to be in line with other dependencies, and
remove a diamond.

Guava is only used in test, explicitly setting the scope to be test.