chore: update maven config for new version of license-maven-plugin
license-maven-plugin added a new licenseSets element to configuration
and deprecated the 3.0 style config. This change updates the config
to use the new style.

See the "NOTE" in  https://github.com/mycila/license-maven-plugin/tree/9a4b59afc6dad5d11a1b321dbe160f16396508e2#detailed-maven-documentation
for the callout of the need for this change.